requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
mpmath==1.3.0
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scikit-commpy==0.8.0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
zipp==3.20.2
